create handler.py and move MessageHandler to it
All handlers are implemented from AbstractHandler
